+++ blueprint/implementations/worker/bullmq/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @forklaunch/implementation-worker-bullmq
 
+## 1.0.26
+
+### Patch Changes
+
+- Internal package updates
+- Updated dependencies
+  - @forklaunch/interfaces-worker@1.0.24
+
 ## 1.0.25
 
 ### Patch Changes